Add tests for the new --delete flag
