+++ test/conftest.py
@@ -25,6 +25,6 @@ def skipif_pre_ansible28(is_pre_ansible28):
         pytest.skip("Valid only on Ansible 2.8+")
 
 
-@pytest.fixture
+@pytest.fixture(scope='session')
 def test_data_dir():
     return os.path.join(os.path.dirname(__file__), 'data')
